Move BernoulliPassRate to punit-core, reuse BinomialProportionEstimator
Reviewer correctly identified an egregious breach of punit's design
principle: statistical calculations belong only in
org.javai.punit.statistics, the dedicated, ArchUnit-isolated home
for the framework's mathematical core. The Wilson-score lower-bound
formula and inverse-normal-CDF approximation I introduced in
punit-api/.../typed/spec/WilsonScore.java duplicated machinery
already implemented and well-tested in
org.javai.punit.statistics.BinomialProportionEstimator.

This commit removes the duplication and the breach:

  - WilsonScore.java + WilsonScoreTest.java deleted from punit-api.
  - BernoulliPassRate moved from punit-api to a new
    org.javai.punit.engine.criteria package in punit-core. The
    Criterion interface stays in punit-api; the implementation lives
    where it can call into the statistics package.
  - BernoulliPassRate.evaluate() empirical path now calls
    BinomialProportionEstimator.lowerBound(successes, trials,
    confidence) — which is also numerically more correct than my
    deleted WilsonScore: the existing estimator uses the one-sided
    z-score (z_{0.95} ≈ 1.645) for a one-sided lower bound; my
    duplicate was using the two-sided z (z_{0.975} ≈ 1.96), giving
    a more conservative bound than the statistics specify.
  - Tests follow the class: BernoulliPassRateTest,
    CriterionResultDetailTest, and InlineSamplingFormTest move to
    punit-core under the new package.
  - Integration tests in punit-core update their imports.
  - TypedEngineArchitectureTest narrows its scope to engine package
    direct (not engine..) — the existing rule wrongly forbade
    engine.criteria from referencing Experiment, but that's a value
    type referenced via the empiricalFrom(Supplier<Experiment>)
    pinning API, not subtype discrimination.

CLAUDE.md gains a §"Statistics isolation rule" enshrining the
principle that motivated this PR's last commit: statistical
calculations live only in org.javai.punit.statistics; types outside
that package consume the existing classes rather than reimplementing
them.
